requirements.txt
accelerate==0.23.0
aiodns==3.0.0
aiofile==3.8.5
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
aniso8601==9.0.1
annotated-types==0.5.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.41.1
bleach==6.0.0
blessed==1.20.0
blinker==1.6.2
cachetools==5.3.0
caio==0.9.12
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.6.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diskcache==5.6.3
dnspython==2.4.2
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.4
Flask==2.3.3
Flask-RESTful==0.3.10
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.17.2
idna==3.4
importlib-metadata==6.6.0
inquirer==3.1.3
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jsonify==0.5
jsonpatch==1.33
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
langchain==0.0.300
langsmith==0.0.40
llama_cpp_python==0.2.6
loguru==0.7.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numexpr==2.8.6
numpy==1.24.3
openai==0.27.4
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pdfkit==1.0.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.4
platformdirs==3.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycares==4.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.6.0
Pympler==1.0.1
pyparsing==3.0.9
pypdf==3.16.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-json-logger==2.0.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
readchar==4.0.5
regex==2023.8.8
replicate==0.13.0
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.4
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.2
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.21
stack-data==0.6.2
stream-chat==4.7.0
streamlit==1.21.0
streamlit-chat==0.0.2.2
sympy==1.12
tenacity==8.2.3
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.2.0.dev20230922
torchaudio==2.2.0.dev20230922
torchvision==0.17.0.dev20230922
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3
uri-template==1.2.0
urllib3==1.26.15
validators==0.20.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.7
widgetsnbextension==4.0.7
xformers==0.0.21
yarl==1.9.1
zipp==3.15.0
